Refactor UserAuthenticationAPIResource into 3 separate resources, Clean HealthResource
