+++ src/main/java/net/whydah/service/Main.java
@@ -1,6 +1,6 @@
 package net.whydah.service;
 
-import net.whydah.service.authapi.UserAuthenticationAPIResource;
+import net.whydah.service.auth.UserAuthenticationResource;
 import net.whydah.service.health.HealthResource;
 import net.whydah.service.proxy.ProxyResource;
 import net.whydah.util.Configuration;
@@ -168,7 +168,7 @@ public class Main {
         // Allow userAuthEndpointConstraintMapping to be accessed without authentication
         ConstraintMapping userAuthEndpointConstraintMapping = new ConstraintMapping();
         userAuthEndpointConstraintMapping.setConstraint(new Constraint(Constraint.NONE, Constraint.ANY_ROLE));
-        userAuthEndpointConstraintMapping.setPathSpec(UserAuthenticationAPIResource.API_PATH + "/*");
+        userAuthEndpointConstraintMapping.setPathSpec(UserAuthenticationResource.API_PATH + "/*");
         securityHandler.addConstraintMapping(userAuthEndpointConstraintMapping);
 
         HashLoginService loginService = new HashLoginService("Whydah-SPAProxyService");
